LawBot_FullStack/requirements.txt
# Backend Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# ML/AI Dependencies
torch>=2.2.0
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
peft>=0.6.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Data Processing
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dateutil>=2.8.2

# Development
pytest>=7.4.3
black>=23.9.1
flake8>=6.1.0
